api/requirements.txt
Flask
Flask-RESTful
PyJWT
pymongo
Jinja2
pytest
pytest-cov
coveralls
faker